Don't show word count history on the stats page, because we don't actually have this information!
The problem is that the first untouched revision is not stored in the database;
when Docstring.source_doc is replaced due to a pull from SVN, word counts
will change. Assuming the first revision was Docstring.source_doc will simply
drastically underestimate the effort people have made.

So, for now, it's better to show only the number of new revisions...