Implement lightweight "OK to apply" review, separate from the main reviews.
In the Numpy Pydocweb, it became obvious that prior to releases we would
need a review mechanism that allows to say "this change to the docstring
is OK to go in SVN". Note that this is different from using the review
system to say "This docstring is Good and Approved"; the OK to Apply
just means the current version is a sufficient improvement over the old one,
and does not introduce regressions.

This commit adds a "OK to Apply" mechanism to Pydocweb. It is easily
accessible for reviewers both on the docstring page, and on 'diff' pages,
which the patch page links to.

DB schema revision is bumped to 5.